split up stats
